Eliminate an script error patching a directory.
The download_and_extract.sh script had a harmless but annonying and
disconcerting error caused by running sed on a directory.

In the patch_cmsis() function a find call was looking for files
matching the pattern *.* which was matching a directory named
.settings in the cmsis sources.  This directory was left over from
running eclipse and shouldn't have been checked in.